Added delay offset code during windtrack load
added spray offset code to run during windtrack load. This is a decent workaround for not having real time offset modification for spray to account for the wake to sleep delay.  Added stopwatch functionality to track when last spraycmd was used but current it is not being used to change offsets.